refactor for subjects lib
